Pol.is login now works in docker in docker

+++ backend/polis-sysbox/startup.sh
@@ -5,16 +5,19 @@ dockerd-entrypoint.sh &
 # Install and configure mkcert for local ssl auth
 WORK_DIR=$(pwd)
 
-curl -JLO "https://dl.filippo.io/mkcert/latest?for=linux/amd64"
-chmod +x mkcert-v*-linux-amd64
-cp mkcert-v*-linux-amd64 /usr/local/bin/mkcert
+if [ ! -f /usr/bin/mkcert ]; then
+    curl -JLO "https://dl.filippo.io/mkcert/latest?for=linux/amd64"
+    chmod +x mkcert-v*-linux-amd64
+    cp mkcert-v*-linux-amd64 /usr/local/bin/mkcert
+fi
 
-mkdir -p /root/.simulacrum/certs
-cd /root/.simulacrum/certs
-mkcert -install   # Created a new local CA at the location returned from mkcert -CAROOT
-mkcert localhost  # Using the local CA at CAROOT, create a new certificate valid for the following names
-
-cd $WORK_DIR 
+if [ ! -d /root/.simulacrum/certs ]; then
+    mkdir -p /root/.simulacrum/certs
+    cd /root/.simulacrum/certs
+    mkcert -install   # Created a new local CA at the location returned from mkcert -CAROOT
+    mkcert localhost  # Using the local CA at CAROOT, create a new certificate valid for the following names
+    cd $WORK_DIR 
+fi
 
 # Wait for docker to be ready
 echo "Waiting for Docker daemon to start..."
@@ -36,7 +39,10 @@ docker info
 if [ ! -f .env ]; then
     echo "Creating .env from example.env"
     cp example.env .env
+    # Suppress auth errors (TODO: FIX)
+    echo "NODE_TLS_REJECT_UNAUTHORIZED=0" >> .env
 fi
+
 
 # Start Polis development environment using docker-compose directly
 echo "Starting Polis development environment..."
